pyproject.toml
[project]
name = "hn-digest"
version = "0.1.0"
description = "Daily Hacker News digest pipeline"
requires-python = ">=3.11"
dependencies = [
    "requests>=2.32",
    "apache-airflow>=2.10",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.9",
    "pre-commit>=4.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=src/hn_digest --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["src/hn_digest"]
